requirements.txt
fastapi
uvicorn
motor
pymongo
ldap3
aio-pika
pydantic-settings
jose
pyjwt
structlog
pytest
pytest-mock
coverage
pytest-asyncio
pydantic[email]